Create Word frequency in String.PY
Word frequency in String
Write a function to get frequency of each word in the sentence arranged in the descending order of the frequency.

Sample Input:

"Python is a high-level, interpreted, general-purpose programming language. Its design philosophy emphasizes code readability with the use of significant indentation. Python is dynamically-typed and garbage-collected."